Cartesian experiment class and an example usage for TestDecodingValidator test class
