+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/io/erasurecode/rawcoder/CartesianProductGenerator.java
@@ -1,0 +1,24 @@
+package org.apache.hadoop.io.erasurecode.rawcoder;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CartesianProductGenerator {
+    public static List<Object[]> generate(Object[][] arrays) {
+        List<Object[]> result = new ArrayList<>();
+        cartesianProductRec(arrays, result, 0, new Object[arrays.length]);
+        return result;
+    }
+
+    private static void cartesianProductRec(Object[][] arrays, List<Object[]> result, int index, Object[] current) {
+        if (index == arrays.length) {
+            result.add(current.clone());
+        } else {
+            for (Object item : arrays[index]) {
+                current[index] = item;
+                cartesianProductRec(arrays, result, index + 1, current);
+            }
+        }
+    }
+
+}
